\name{hymod.sim}
\alias{hymod.sim}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Function Software working in the R environment for producing confidence limits of a generic prediction through comparison with observed data. Confidence limits are estimated by applying the BLUECAT method that transforms any deterministic model in a stochastic model, from which mean stochastic simulation and confidence limits are obtained. For more details please refer to https:\www.albertomontanari.it\bluecat and Koutsoyiannis and Montanari (2022).}
\description{
The function takes as input a calibration data set formed by deterministic model predicitons and corresponding observed values. By comparing prediction with observation BLUECAT corrects the prediction itself and computes confidence bands. The function requires the libraries "devtools", "DescTools" and "DEoptim".}
\usage{
bluecat.sim<-function(resultcalib,dmodelsim,predsmodel="avg",empquant=F,siglev=0.2,m=100,m1=80,qoss=NULL,plotflag=F,cpptresh=0)}
\arguments{
\item{resultcalib}{It is a list with elements $qsim and qoss}
\item{dmodelsim}{Vector of deterministic model prediction whose uncertainty is to be estimated.}
\item{predsmodel}{Character variable to specify if stochastic prediction is generated by the mean (predsmodel="avg") or the median (predsmodel="mdn") of the conditional distribution. Default is predsmodel="avg".}
\item{empquant}{Logical variable to specify if empirical quantiles or robust estimation is to be used for estimating confidence limits. Default is robust estimation, namely, empquant=F.}
\item{siglev}{Significance level for confidence limits estimation. Default is siglev=0.2}
\item{m}{Parameter to determine the sample size of river flow neighbours to be used for estimating the probability distribution of true river flow conditioned to the simulated river flow. Default is m=100.}
\item{m1}{Number of k-moments used to estimate the PBF distribution on the sample of mean stochastic prediction to make robust quantile estimation. Default is m1=80.}
\item{paramd}{Initial parameter values for the PBF distribution fitting the sample of mean stochastic prediction to make robust quantile estimation. Default is paramd=c(0.1,1,10,NA).}
\item{lowparamd}{Lower bound for the parameter values of the PBF distribution fitting the sample of mean stochastic prediction to make robust quantile estimation. Default is lowparamd=c(0.001,0.001,0.001,0).}
\item{upparamd}{Upper bound for the parameter values of the PBF distribution fitting the sample of mean stochastic prediction to make robust quantile estimation. Default is upparamd=c(1,5,20,NA).}
\item{qoss}{Vector of observed data corresponding to the model prediction whose uncertainty is to be estimated if available. It is not necessary for uncertainty estimation. If provided, it must have the same length as the vector of modelprediction dmodelsim.}
\item{plotflg}{Logical value. Specifies if diagnostic is to be returned. Default is plot=F. If plot=T a scapperplot of observed versus simulated by the stochastic model river flows is returned along with its efficiency. The diagnostic plots are also returned, along with the percentage of observations lying outside the confidence bands.}
\item{cpptresh}{Low flow threshold to draw the diagnostic plots and to compute percentage of points lying outside the confidence band. Default value is cpptresh=0, which means that zero values for the simulated river flow are not considered when drawing the plots and computing percentages.}
}

\value{If eff=T a single numerical correspoding to the Nash-Sutcliffe (1968) efficiency is returned if qoss is given (otherwise efficiency cannot be computed and therefore is not returned).
A list with the following items is returned:
\item{outputlist$stochprediction}{River flow obtained as mean stochastic simulation.}
\item{outputlist$lowlimit}{Lower confidence limit obtained with BLUECAT.}
\item{outputlist$uplimit}{Upper confidence limit obtained with BLUECAT.}
\item{effdmodel}{The Nash Sutcliffe efficiency of the simulation by the deterministic model. This is returned only if qoss is given.}

\details{
Uncertainty is estimated by applying BLUECAT, "A Brisk Local Uncertainty Estimator for Hydrologic Simulations and Predictions", introduced by Koutsoyiannis and Montanari (2021). BLUECAT estimates the lower and upper confidence limits of the HyMod simulation for a given confidence level. Goodness of fit testing for BLUECAT is performed by plotting a scatterplot of observed river flow against the mean stochastic simulation obtained with BLUECAT along with the related confidence limits, by plotting the Coverage Probability Plot and computing the Prediction Technical Performance indexes as in Koutsoyiannis and Montanari (2022).}

\author{Demetris Koutsoyiannis and Alberto Montanari, email: alberto.montanari@unibo.it.}

\references{
Koutsoyiannis, D., & Montanari, A. (2022). Bluecat: A local uncertainty estimator for deterministic simulations and predictions. Water Resources Research, 58(1), e2021WR031215.

Nash, J. E., and Sutcliffe, J. V., (1970), River flow forecasting through conceptual models part I – A discussion of principles, J. Hydrol., 10, 282–290, \doi{https://doi.org/10.1016/0022-1694(70)90255-6}. 
}

\examples{
## Load data
data(arnosubbiano)

## Model parameterisation with hymod.par
pr1=hymod.par(c(100,1,0.5,200,0.5),area=752,tdelta=86400,e=arnosubbiano[,3][1:7305],p=arnosubbiano[,2][1:7305],nstep=length(p[1:7305]),qoss=arnosubbiano[,4][1:7305],qinitial=15,lower=c(10,0.1,0.1,0.1,0.1),upper=c(800,10,0.9,1000,100),opt="DEoptim")

## Run BLUECAT to perform simulation in calibration mode
pr2=hymod.sim(pr1$par,area=752,tdelta=86400,e=arnosubbiano[,3][1:7305],p=arnosubbiano[,2][1:7305],qinitial=15,qoss=arnosubbiano[,4][1:7305],resultcalib=pr1,bluecat=T,empquant=F,plot=T)

## Run BLUECAT to perform simulation in validation mode
pr3=hymod.sim(pr1$par,area=752,tdelta=86400,e=arnosubbiano[,3][7306:8036],p=arnosubbiano[,2][7306:8036],qinitial=15,qoss=arnosubbiano[,4][7306:8036],resultcalib=pr1,bluecat=T,empquant=F,plot=T)

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{Hydrology}
\keyword{BLUECAT}
\keyword{Uncertainty}% __ONLY ONE__ keyword per line
